seer/requirements.txt
# temp dependencies for pysage3
httpx
rectpack
networkx
graphviz
pydantic>=2
websocket-client
redis
requests
ws4py
celery~=5.2.2
requests>=2.32.0 # not directly required, pinned by Snyk to avoid a vulnerability
numpy>=1.22.2 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability

# SAGE3
git+https://github.com/SAGE-3/next.git@dev-langchain-deploy#subdirectory=foresight
python-dotenv

# Web API
fastapi==0.111.0
fastapi-cli
uvicorn

# data models
pydantic>=2.0.0

# AI with langchain
langchain==0.2.3
langchain-openai==0.1.8
langchain-huggingface==0.0.3
langchain-nvidia-ai-endpoints==0.2.0

# Image
PIL

# Web scrapping
playwright
